Remove react router and routes
