+++ user_interfaces/gui/mdm-react/package.json
@@ -12,7 +12,6 @@
     "react-bootstrap": "^0.32.4",
     "react-dom": "^16.5.1",
     "react-redux": "^5.0.7",
-    "react-router-dom": "^4.3.1",
     "react-scripts": "^1.1.5",
     "react-sigma": "^1.2.30",
     "react-virtualized": "^9.20.1",
