+++ user_interfaces/gui/mdm-react/src/components/AppHeader/app-header.css
@@ -18,7 +18,3 @@
 .mdm-header .subtitle {
     margin: 0;
 }
-
-.mdm-header .title:hover {
-    cursor: pointer;
-}
