--- user_interfaces/gui/mdm-react/src/components/AppNavigation/representation/AppNavLink/index.jsx
@@ -1,21 +1,0 @@
-import React, { Fragment } from 'react';
-import PropTypes from 'prop-types';
-
-import { MDMLink } from '../../..';
-
-const AppNavLink = ({ icon, label }) => (
-  <Fragment>
-    <div className="nav-icon">
-      <i className="material-icons">{icon}</i>
-    </div>
-    <MDMLink label={label} />
-  </Fragment>
-);
-
-AppNavLink.propTypes = {
-  icon: PropTypes.string.isRequired,
-  label: PropTypes.string.isRequired,
-};
-AppNavLink.defaultProps = {};
-
-export default AppNavLink;
