--- user_interfaces/gui/mdm-react/src/components/AppNavigation/representation/index.js
@@ -1,5 +1,0 @@
-import AppNavLink from './AppNavLink/index';
-
-export {
-  AppNavLink,
-};
