+++ user_interfaces/gui/mdm-react/src/components/CreationButton/representation/CreationPanel/DeploymentButton.jsx
@@ -7,7 +7,6 @@ class DeploymentButton extends Component {
   render() {
     const {
       dataGathered,
-      deployed,
       deploying,
       gatheringData,
       isSystemActive,
@@ -52,7 +51,6 @@ class DeploymentButton extends Component {
 
 DeploymentButton.propTypes = {
   dataGathered: PropTypes.bool.isRequired,
-  deployed: PropTypes.bool.isRequired,
   deploying: PropTypes.bool.isRequired,
   gatheringData: PropTypes.bool.isRequired,
   isSystemActive: PropTypes.bool.isRequired,
@@ -65,7 +63,6 @@ function mapStateToProps({
   graph: {
     deployment: {
       dataGathered,
-      deployed,
       deploying,
       gatheringData,
     },
@@ -73,7 +70,6 @@ function mapStateToProps({
 }) {
   return {
     dataGathered,
-    deployed,
     deploying,
     gatheringData,
   };
