--- user_interfaces/gui/mdm-react/src/components/Link/index.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-import './link.css';
-
-const MDMLink = ({ label }) => (
-  <span className="mdm-link">
-    {label}
-  </span>
-);
-
-MDMLink.propTypes = {
-  label: PropTypes.string.isRequired,
-};
-MDMLink.defaultProps = {};
-
-export default MDMLink;
