--- user_interfaces/gui/mdm-react/src/components/Link/link.css
@@ -1,6 +1,0 @@
-.mdm-link {
-    text-decoration: none;
-    text-transform: capitalize;
-    color: aliceblue;
-    font-weight: bold;
-}
