+++ user_interfaces/gui/mdm-react/src/providers/route/index.jsx
@@ -1,14 +1,11 @@
 import React, { Component } from 'react';
-import { BrowserRouter } from 'react-router-dom';
 
 import App from '../../routes/index';
 
 class AppRouter extends Component {
   render() {
     return (
-      <BrowserRouter>
-        <App />
-      </BrowserRouter>
+      <App />
     );
   }
 }
