--- user_interfaces/gui/mdm-react/src/routes/deploy/index.jsx
@@ -1,14 +1,0 @@
-import React, { Component } from 'react';
-
-class DeployPage extends Component {
-  render() {
-    return (
-      <div>Deploy</div>
-    );
-  }
-}
-
-DeployPage.propTypes = {};
-DeployPage.defaultProps = {};
-
-export default DeployPage;
