--- user_interfaces/gui/mdm-react/src/routes/monitor/index.jsx
@@ -1,14 +1,0 @@
-import React, { Component } from 'react';
-
-class MonitorPage extends Component {
-  render() {
-    return (
-      <div>Monitor</div>
-    );
-  }
-}
-
-MonitorPage.propTypes = {};
-MonitorPage.defaultProps = {};
-
-export default MonitorPage;
