--- user_interfaces/gui/mdm-react/src/routes/welcome/welcome.css
@@ -1,1 +1,0 @@
-
